pyproject.toml
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "spark-vtuber"
version = "0.1.0"
description = "AI-powered VTuber streaming system for NVIDIA DGX Spark"
authors = ["Jack Hacksman <slack@hannis.io>"]
readme = "README.md"
license = "MIT"
packages = [{include = "spark_vtuber", from = "src"}]

[tool.poetry.dependencies]
python = "^3.10"

# LLM Inference
vllm = "^0.4.0"
transformers = "^4.40.0"
peft = "^0.10.0"
accelerate = "^0.28.0"
torch = "^2.2.0"
bitsandbytes = "^0.43.0"

# TTS
TTS = "^0.22.0"
soundfile = "^0.12.0"
scipy = "^1.12.0"

# STT
faster-whisper = "^1.0.0"
pyannote-audio = "^3.1.0"
webrtcvad = "^2.0.10"

# Memory/RAG
chromadb = "^0.4.0"
sentence-transformers = "^2.5.0"

# Avatar
websockets = "^12.0"
librosa = "^0.10.0"
numpy = "^1.26.0"

# Chat Integration
irc = "^20.0"
aiohttp = "^3.9.0"

# Game Integration
pyautogui = "^0.9.0"
pillow = "^10.2.0"
opencv-python = "^4.9.0"

# Core
pydantic = "^2.6.0"
pydantic-settings = "^2.2.0"
loguru = "^0.7.0"
typer = "^0.9.0"
rich = "^13.0.0"
python-dotenv = "^1.0.0"
asyncio-mqtt = "^0.16.0"
aiofiles = "^23.2.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-asyncio = "^0.23.0"
pytest-cov = "^4.1.0"
black = "^24.0.0"
ruff = "^0.3.0"
mypy = "^1.9.0"
pre-commit = "^3.6.0"

[tool.poetry.scripts]
spark-vtuber = "spark_vtuber.main:app"

[tool.black]
line-length = 100
target-version = ['py310']

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-v --cov=spark_vtuber --cov-report=term-missing"
